requirements.txt
Django==3.0.8
django-allauth==0.42.0
django-sortedm2m==3.0.2
Pillow==7.2.0
django-heroku==0.3.1